Moved functions orbit and orbit_transversal from perm_groups.py to perm_algorithms.py
