+++ sympy/combinatorics/perm_algorithms.py
@@ -1,5 +1,5 @@
-from sympy.combinatorics.permutations import Permutation, perm_af_mul, perm_af_muln, perm_af_invert
-from sympy.combinatorics.perm_groups import PermutationGroup, orbit, orbit_transversal
+from sympy.combinatorics.permutations import Permutation, perm_af_mul, perm_af_muln, perm_af_invert, _new_from_array_form
+from sympy.combinatorics.perm_groups import PermutationGroup
 
 """
     References for tensor canonicalization:
@@ -18,6 +18,116 @@ from sympy.combinatorics.perm_groups import PermutationGroup, orbit, orbit_trans
     [4] xperm.c part of XPerm written by J. M. Martin-Garcia
         http://www.xact.es/index.html
 """
+
+def orbit(genv, alpha):
+    r"""
+    Compute the orbit of alpha `\{g(\alpha) | g \in G\}` as a set.
+
+    genv   generators of the Group G in array form
+    alpha  starting point of the orbit
+
+    The time complexity of the algorithm used here is `O(|Orb|*r)` where
+    `|Orb|` is the size of the orbit and `r` is the number of generators
+    of the group. For a proof of correctness, see [1], p.78.
+
+    Examples
+    ========
+
+    >>> from sympy.combinatorics.perm_algorithms import orbit
+    >>> from sympy.combinatorics.permutations import Permutation
+    >>> genv = [[1,2,0,4,5,6,3]]
+    >>> orbit(genv, 0)
+    set([0, 1, 2])
+    >>> orbit(genv, 4)
+    set([3, 4, 5, 6])
+
+    See Also
+    ========
+
+    orbit_transversal
+
+    References
+    ==========
+
+    [1] Holt, D., Eick, B., O'Brien, E.
+    "Handbook of computational group theory"
+
+    """
+    n = len(genv[0])
+    orb = [alpha]
+    used = [False]*n
+    used[alpha] = True
+    for b in orb:
+        for gen in genv:
+            temp = gen[b]
+            if used[temp] == False:
+                orb.append(temp)
+                used[temp] = True
+    return set(orb)
+
+def orbit_transversal(genv, alpha, pairs=False, af=False):
+    r"""
+    Computes a transversal for the orbit of ``alpha`` as a set.
+
+    genv   generators of the Group G
+    alpha  starting point of the orbit
+    pairs  = True return the list of pairs
+           `(\beta, g_\beta)`. For a proof of correctness, see [1], p.79
+    af     = True genv and the transversal elements are in array form
+
+    For a permutation group `G`, a transversal for the orbit
+    `Orb = \{g(\alpha) | g \in G\}` is a set
+    `\{g_\beta | g_\beta(\alpha) = \beta\}` for `\beta \in Orb`.
+    Note that there may be more than one possible transversal.
+
+    Examples
+    ========
+    >>> from sympy.combinatorics.permutations import _new_from_array_form
+    >>> from sympy.combinatorics.named_groups import DihedralGroup
+    >>> from sympy.combinatorics.perm_algorithms import orbit_transversal
+    >>> G = DihedralGroup(6)
+    >>> orbit_transversal(G.generators, 0)
+    [Permutation([0, 1, 2, 3, 4, 5]), Permutation([1, 2, 3, 4, 5, 0]),
+    Permutation([5, 4, 3, 2, 1, 0]), Permutation([2, 3, 4, 5, 0, 1]),
+    Permutation([4, 3, 2, 1, 0, 5]), Permutation([3, 4, 5, 0, 1, 2])]
+
+    See Also
+    ========
+
+    orbit
+
+    References
+    ==========
+
+    [1] Holt, D., Eick, B., O'Brien, E.
+    "Handbook of computational group theory"
+
+    """
+    if af:
+        n = len(genv[0])
+    else:
+        n = genv[0].size
+    tr = [(alpha, range(n))]
+    used = [False]*n
+    used[alpha] = True
+    if not af:
+        genv = [h.array_form for h in genv]
+    for pair in tr:
+        for gen in genv:
+            temp = gen[pair[0]]
+            if used[temp] == False:
+                tr.append((temp, perm_af_mul(gen, pair[1])))
+                used[temp] = True
+    if pairs:
+        if af:
+            return tr
+        else:
+            tr = [(v, _new_from_array_form(h)) for v, h in tr]
+            return tr
+    if af:
+        return [pair[1] for pair in tr]
+    else:
+        return [_new_from_array_form(pair[1]) for pair in tr]
 
 def dummy_sgs(d, sym, n):
     """
